+++ Tron/Architect.API.Tron.Contracts/Cotizacion/MapfreMas.cs
@@ -116,7 +116,7 @@ namespace Architect.API.Tron.Contracts.Cotizacion
         /// <summary>
         /// Valor del vehículo asegurado.
         /// </summary>
-        [DataMember(), JsonProperty()] public int IMP_VR { get; set; }
+        [DataMember(), JsonProperty()] public double IMP_VR { get; set; }
         /// <summary>
         /// ¿Es cliente nuevo?.
         /// </summary>
